pyproject.toml
[project]
name = "dynamic-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "graphviz",
    "httpx",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langgraph>=0.6.3",
    "matplotlib",
    "networkx",
    "pillow",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "rich",
    "tavily-python>=0.7.10",
    "typer>=0.16.0",
    "typing-extensions",
]
